requirements.txt
google-genai==0.3.0
gradio==5.45.0
requests==2.32.3
python-dotenv==1.0.1
feedparser==6.0.11
trafilatura==1.12.2
dateparser==1.2.0
